Basic admin role

+++ server/controllers/users.js
@@ -8,11 +8,15 @@ module.exports = class {
    * @param {*} res 
    */
   findAll(req, res) {
-    models.user.findAll().then((users) => {
-      res.json(users);
-    }).catch((err) => {
-      res.status(500).send(err.errors);
-    });
+    if (!req.user.isAdmin) {
+      res.sendStatus(403);
+    } else {
+      models.user.findAll().then((users) => {
+        res.json(users);
+      }).catch((err) => {
+        res.status(500).send(err.errors);
+      });
+    }
   }
 
   /**
@@ -21,15 +25,19 @@ module.exports = class {
    * @param {*} res 
    */
   findOne(req, res) {
-    models.user.findById(req.params.id).then((users) => {
-      if (users) {
-        res.json(users);
-      } else {
-        res.sendStatus(404);
-      }
-    }).catch((err) => {
-      res.status(500).send(err.errors);
-    });
+    if ((req.user.id !== req.params.id) && !req.user.isAdmin) {
+      res.sendStatus(403);
+    } else {
+      models.user.findById(req.params.id).then((users) => {
+        if (users) {
+          res.json(users);
+        } else {
+          res.sendStatus(404);
+        }
+      }).catch((err) => {
+        res.status(500).send(err.errors);
+      });
+    }
   }
 
   /**
@@ -51,15 +59,19 @@ module.exports = class {
    * @param {*} res 
    */
   update(req, res) {
-    models.user.update(req.body, {
-      where: {
-        id: req.params.id
-      }
-    }).spread((count, users) => {
-      res.sendStatus(count ? 200: 404);
-    }).catch((err) => {
-      res.status(500).send(err.errors);
-    });
+    if ((req.user.id !== req.params.id) && !req.user.isAdmin) {
+      res.sendStatus(403);
+    } else {
+      models.user.update(req.body, {
+        where: {
+          id: req.params.id
+        }
+      }).spread((count, users) => {
+        res.sendStatus(count ? 200 : 404);
+      }).catch((err) => {
+        res.status(500).send(err.errors);
+      });
+    }
   }
 
   /**
@@ -68,14 +80,18 @@ module.exports = class {
    * @param {*} res 
    */
   delete(req, res) {
-    models.user.destroy({
-      where: {
-        id: req.params.id
-      }
-    }).then((count) => {
-      res.sendStatus(count === 1 ? 200 : 404);
-    }).catch((err) => {
-      res.status(500).send(err.errors);
-    });
+    if (!req.user.isAdmin) {
+      res.sendStatus(403);
+    } else {
+      models.user.destroy({
+        where: {
+          id: req.params.id
+        }
+      }).then((count) => {
+        res.sendStatus(count === 1 ? 200 : 404);
+      }).catch((err) => {
+        res.status(500).send(err.errors);
+      });
+    }
   }
 };
